+++ app/components/decor/suite/forms/checkbox.rb
@@ -132,7 +132,7 @@ module Decor
           end
 
           if !floating_error_text? && errors?
-            p(class: "decor:suite-field-help decor:text-suite-danger-700 #{margin}") do
+            p(class: "decor:suite-field-help decor:text-suite-danger-500 #{margin}") do
               plain error_text
             end
           end