+++ app/components/decor/suite/forms/expanding_checkbox_collection.rb
@@ -86,7 +86,7 @@ module Decor
           end
 
           if !floating_error_text? && errors?
-            p(class: "decor:suite-field-help decor:text-suite-danger-700") { plain error_text }
+            p(class: "decor:suite-field-help decor:text-suite-danger-500") { plain error_text }
           end
         end
       end